Add TLS support (and enable HTTP/2)
As with the root path, the keypair files are watched for changes and
reloaded automatically.  If the files are symbolic links, only changes
to the links themselves, not their targets, are monitored; therefore,
either the keypair are links to immutable files (like the root path is)
or they are regular files that are written to when the keypair changes.