Lock NixStorePathServer on ServeHTTP
h.resolvedRoot and h.etag must match one another.  When writing,
a lock is already taken for that purpose, but it was being ignored
when reading.

We might want to move actual serving elsewhere to allow the unlock to
happen earlier, but for now I'll just note it down so I know where to
look if it slows down too much.