Read site-specific headers from /.snowweb/headers
Provide a mechanism for sites to specify additional headers
to be sent with every response.  This decouples headers like
Content-Security-Policy from the server config and allows them
to evolve alongside the websites themselves.

Site-specific headers are read from `.snowweb/headers` in the
derivation's output path.  This file has the same format as HTTP
headers themselves, a series of `Key: Value` lines followed by a
blank line.  (Per the POSIX definition of a text line, the file
must end with the two bytes `0a 0a`.)