Verify TLS client certificate if possible
mTLS seems to be the better way to authenticate clients for remote API
access (at least from Lorkep's POV, where everything uses mTLS).  The
first step in that is allowing the user to provide a CA bundle to verify
against and tell Go's TLS stack to verify the client certificate *if*
one is given.  Authenticated endpoint handlers can then look into the
request to see if client authentication was performed.