Redirect the "log" standard logger to use zerolog
While Go's standard log package is not used in SnowWeb directly, it is
still used by the rest of the standard library, like in http.Server, and
even though the latter lets one specify an alternative Logger instance
(as the ErrorLog field) it's still bound the log package because Logger
is a struct, not an interface other packages can fulfill (see
github:golang/go#13182 for a proposal to convert it into an interface).

The end result is that we have two competing logging solutions
throughout the program, writing to possibly different places in
definitively different formats.

Luckily for us, Logger amounts to a glorified Printf and there is no
metadata (like logging level) to lose if we treat it as such, so we can
just have it use zerolog's logger as its output stream.