Stop watching TLS keypair for changes
Since switching focus to serve Nix store paths, filesystem monitoring
has become unnecessary.  We dropped the root path watcher at the switch
and now we're removing the watchers for the TLS certificate and key
files.

We still need a way to reload the keypair on demand, and we'll be
addressing that in the next commit.