Fix detection of directory requests
While investigating how to drop the golang.org/x/sys/unix package and
whether fs.Open would return EISDIR on Windows, I found out that it does
not return EISDIR at all.  What does return EISDIR is reading from the
file descriptor.

Since we don't read from the file until it's too late, what we need to
do instead is open it and then stat it to check if it's a directory.