Fix systemd socket logic
SD_LISTEN_FDS_START is not an environment variable, but a constant.
I should read the documentation more carefully next time.

Instead of fixing the parse logic, I decided to bite the bullet and
pull a dependency to do it for me.  github.com/coreos/go-systemd also
provides functions to handle service notifications and journal writing,
which I can make used of in the future.