Fix service command line
Due to github:alecthomas/kong#153, $SNOWWEB_INSTALLABLE is not used
currently so we need to specify it in the command line regardless.
While we're at it, take advantage of freeform modules' partial typing
to ensure the installable is given and to provide defaults to ACME
options (but don't try to perform exhaustive checking).